# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vokal Director API",
    "description": "Контракт между фронтендом и сервером Vokal Director.\n\n**Что здесь важно знать до чтения адресов.**\n\n* Большая часть адресов отвечает `501`. Это не заготовка на будущее: схема\n  запроса и ответа настоящая и полная, реализации за адресом еще нет. В\n  `details.missing` сказано, чего именно не хватает. Заглушек, возвращающих\n  правдоподобные данные, в API нет — мок в ответе сервера хуже отказа.\n* Ошибка всегда выглядит одинаково: `{\"error\": {\"code\", \"message\",\n  \"requestId\", \"details\"}}`.\n* Имена полей — camelCase. Идентификаторы — непрозрачные строки.\n* Незнакомое поле в теле запроса — ошибка `422`, а не тихо отброшенное\n  значение.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects/{project_id}/director/suggestions": {
      "get": {
        "tags": [
          "директор"
        ],
        "summary": "Предложения директора",
        "operationId": "list_suggestions_api_projects__project_id__director_suggestions_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectorSuggestionListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/director/actions": {
      "post": {
        "tags": [
          "директор"
        ],
        "summary": "Выполнить одно действие директора",
        "description": "Создает новую версию аранжировки и помечает материалы, которые устарели.",
        "operationId": "apply_action_api_projects__project_id__director_actions_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectorActionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectorActionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Состояние объекта изменилось: обновите данные и повторите.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Данные запроса не прошли проверку схемы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/director/actions/batch": {
      "post": {
        "tags": [
          "директор"
        ],
        "summary": "Собрать одну версию сразу из нескольких действий",
        "description": "Действия применяются в указанном порядке и дают одну версию с общим перечнем изменений, а не цепочку из промежуточных версий.",
        "operationId": "apply_actions_batch_api_projects__project_id__director_actions_batch_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectorActionBatchRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectorActionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Состояние объекта изменилось: обновите данные и повторите.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Данные запроса не прошли проверку схемы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/director/chat": {
      "post": {
        "tags": [
          "директор"
        ],
        "summary": "Реплика директору",
        "description": "Свободный текст превращается в действие из закрытого списка. Непонятая команда — честный ответ «не понял», а не выполнение похожего действия.",
        "operationId": "chat_api_projects__project_id__director_chat_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "Идентификатор проекта",
              "title": "Project Id"
            },
            "description": "Идентификатор проекта"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectorChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectorChatResponse"
                }
              }
            }
          },
          "401": {
            "description": "Нужен вход в аккаунт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Доступ к этому объекту закрыт.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден или уже удален.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Данные запроса не прошли проверку схемы.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Слишком много запросов подряд.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "501": {
            "description": "Адрес объявлен, реализации еще нет. В details.missing сказано, чего не хватает.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArrangementVersionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "kind": {
            "$ref": "#/components/schemas/VersionKind"
          },
          "parentVersionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parentversionid",
            "description": "Родительская версия. Пусто у исходной."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "createdBy": {
            "type": "string",
            "title": "Createdby",
            "description": "Кто создал: человек или действие директора"
          },
          "status": {
            "$ref": "#/components/schemas/VersionStatus"
          },
          "changes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changes",
            "description": "Что изменилось относительно родителя"
          },
          "artifactsSnapshot": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ArtifactOut"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifactssnapshot",
            "description": "Материалы, с которыми создана версия. По ним делается откат. Пусто — материалы этой версии неизвестны, и откат пометит перенесенные к пересборке."
          },
          "isCurrent": {
            "type": "boolean",
            "title": "Iscurrent",
            "description": "Выбрана ли версия сейчас"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "label",
          "kind",
          "createdAt",
          "createdBy",
          "status",
          "changes",
          "isCurrent"
        ],
        "title": "ArrangementVersionOut",
        "description": "Версия аранжировки."
      },
...
      "ArtifactType": {
        "type": "string",
        "enum": [
          "score",
          "part",
          "tab",
          "chords",
          "lyrics",
          "midi",
          "stem",
          "minus",
          "click",
          "practice",
          "teacher",
          "student",
          "zip"
        ],
        "title": "ArtifactType"
      },
...
      "ChatMessageOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "author": {
            "$ref": "#/components/schemas/ChatAuthor"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "author",
          "text",
          "createdAt"
        ],
        "title": "ChatMessageOut"
      },
...
      "DirectorActionBatchRequest": {
        "properties": {
          "actionIds": {
            "items": {
              "$ref": "#/components/schemas/DirectorActionId"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Actionids",
            "description": "Порядок значим: действия применяются в нем"
          },
          "baseVersionId": {
            "type": "string",
            "title": "Baseversionid"
          },
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "Подпись будущей версии"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "actionIds",
          "baseVersionId"
        ],
        "title": "DirectorActionBatchRequest",
        "description": "Сборка одной версии сразу из нескольких действий."
      },
      "DirectorActionId": {
        "type": "string",
        "enum": [
          "transpose-down-2",
          "merge-guitars",
          "move-strings-to-keys",
          "simplify-drums",
          "beginner-bass",
          "boost-chorus",
          "practice-without-bass",
          "education-version",
          "advanced-student-part",
          "student-ensemble",
          "lesson-analysis"
        ],
        "title": "DirectorActionId",
        "description": "Детерминированная команда директора.\n\nСписок закрытый: LLM выбирает действие из него, а не правит материалы\nнапрямую. Незнакомый код — `422`, а не попытка угадать намерение."
      },
      "DirectorActionRequest": {
        "properties": {
          "actionId": {
            "$ref": "#/components/schemas/DirectorActionId"
          },
          "baseVersionId": {
            "type": "string",
            "title": "Baseversionid",
            "description": "Версия, к которой применяется действие. Обязательна: без нее правка легла бы на версию, которую пользователь уже не видит на экране."
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "actionId",
          "baseVersionId"
        ],
        "title": "DirectorActionRequest",
        "description": "Одно действие директора."
      },
      "DirectorActionResponse": {
        "properties": {
          "version": {
            "$ref": "#/components/schemas/ArrangementVersionOut"
          },
          "result": {
            "$ref": "#/components/schemas/DirectorActionResultOut"
          },
          "appliedActionIds": {
            "items": {
              "$ref": "#/components/schemas/DirectorActionId"
            },
            "type": "array",
            "title": "Appliedactionids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "result",
          "appliedActionIds"
        ],
        "title": "DirectorActionResponse",
        "description": "Ответ на действие: новая версия и перечень последствий."
      },
      "DirectorActionResultOut": {
        "properties": {
          "versionLabel": {
            "type": "string",
            "title": "Versionlabel"
          },
          "versionKind": {
            "$ref": "#/components/schemas/VersionKind"
          },
          "historyTitle": {
            "type": "string",
            "title": "Historytitle",
            "description": "Строка для истории изменений"
          },
          "changes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changes"
          },
          "staleArtifactTypes": {
            "items": {
              "$ref": "#/components/schemas/ArtifactType"
            },
            "type": "array",
            "title": "Staleartifacttypes",
            "description": "Типы материалов, которые придется пересобрать после этого действия"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "versionLabel",
          "versionKind",
          "historyTitle",
          "changes",
          "staleArtifactTypes"
        ],
        "title": "DirectorActionResultOut",
        "description": "Что сделало действие. Повторяет `DirectorActionResult` из `types.ts`."
      },
      "DirectorChatRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "DirectorChatRequest",
        "description": "Реплика человека директору."
      },
      "DirectorChatResponse": {
        "properties": {
          "message": {
            "$ref": "#/components/schemas/ChatMessageOut"
          },
          "reply": {
            "$ref": "#/components/schemas/ChatMessageOut"
          },
          "suggestedActionId": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectorActionId"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message",
          "reply"
        ],
        "title": "DirectorChatResponse",
        "description": "Ответ директора.\n\n`suggested_action_id` пустой — нормальный исход: команду не поняли. Честный\nответ «не понял» лучше, чем выполнение похожего действия."
      },
      "DirectorSuggestionImpact": {
        "type": "string",
        "enum": [
          "arrangement",
          "education",
          "export",
          "review"
        ],
        "title": "DirectorSuggestionImpact"
      },
      "DirectorSuggestionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DirectorSuggestionOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DirectorSuggestionListResponse"
      },
      "DirectorSuggestionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "actionId": {
            "$ref": "#/components/schemas/DirectorActionId"
          },
          "scenario": {
            "$ref": "#/components/schemas/DirectorSuggestionScope"
          },
          "impact": {
            "$ref": "#/components/schemas/DirectorSuggestionImpact"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "actionId",
          "scenario",
          "impact"
        ],
        "title": "DirectorSuggestionOut",
        "description": "Предложение директора: описание и код действия за ним."
      },
      "DirectorSuggestionScope": {
        "type": "string",
        "enum": [
          "band",
          "education",
          "both"
        ],
        "title": "DirectorSuggestionScope"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Машинный код для ветвления на клиенте"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Текст для показа пользователю"
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid",
            "description": "Идентификатор запроса: по нему поддержка ищет след в логах"
          },
          "details": {
            "title": "Details",
            "description": "Подробности, зависят от кода. Для validation_error — {fields: [{field, reason}]}, для not_implemented — {endpoint, missing, docs}."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "description": "Тело ошибки. Форма задана `app/core/errors.py` и одинакова для всего API."
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Конверт ошибки: любой ответ с кодом 4xx/5xx выглядит так."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionKind": {
        "type": "string",
        "enum": [
          "original",
          "band",
          "easy",
          "original-like",
          "advanced",
          "concert",
          "ensemble",
          "after-rehearsal",
          "after-lesson"
        ],
        "title": "VersionKind"
      },
...
      "VersionStatus": {
        "type": "string",
        "enum": [
          "draft",
          "needs_review",
          "approved",
          "distributed"
        ],
        "title": "VersionStatus"
      }
    }
  }
}